Refactor OFW message processing and statistics accumulation
- Moved `parseMessage` and `processMessages` functions to `utils/ofw/parser.js`.
- Introduced `accumulateStats` function in `utils/ofw/stats.js` for better statistics handling.
- Updated `writeFile` and `parsePdf` functions to `utils/fs.js` and `utils/pdf.js`, respectively.
- Adjusted imports and exports accordingly in `index.js` and `utils.js`.
- Simplified output functions in `outputMarkdownSummary` and `outputCSV` to utilize new formatting functions.
