+++ utils.js
@@ -1,5 +1,5 @@
 const fs = require("fs");
-const pdf = require("pdf-parse");
+const { parsePdf } = require("./utils/pdf");
 const {
     getWeekString,
     parseDate,
@@ -22,21 +22,7 @@ function writeFile(filePath, data) {
     }
 }
 
-/**
- * Parses the PDF file at the specified file path.
- * @param {string} filePath - The path to the PDF file.
- * @return {Promise<string>} - A promise that resolves to the text content of the PDF file.
- */
-async function parsePdf(filePath) {
-    try {
-        const dataBuffer = fs.readFileSync(filePath);
-        const data = await pdf(dataBuffer);
-        return data.text;
-    } catch (error) {
-        console.error(`Failed to parse PDF at ${filePath}:`, error);
-        throw error; // Re-throw the error to be caught by the calling function
-    }
-}
+// parsePdf moved to `utils/pdf.js`
 
 exports.getWeekString = getWeekString;
 exports.parsePdf = parsePdf;
